requirements.txt
python-telegram-bot==20.8
openai==1.12.0
anthropic==0.18.1
requests==2.31.0
python-dotenv==1.0.1
Pillow==10.2.0
numpy==1.26.4
moviepy==1.0.3
pydub==0.25.1
SQLAlchemy==2.0.27
psutil==5.9.8
googlesearch-python==1.2.3
beautifulsoup4==4.12.3
webdriver-manager==4.0.1
selenium==4.18.1
aiohttp==3.9.3
fastapi==0.109.2
uvicorn==0.27.1
jinja2==3.1.3
pyyaml==6.0.1
